pyproject.toml
[tool.poetry]
name = "glowing-telegram"
version = "0.1.0"
description = "Telegram bots for life and leisure"
authors = ["Abhishek Juneja <abhishek.juneja145@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
black = "^21.9b0"
python-telegram-bot = "^13.7"
ipython = "^7.28.0"
bpython = "^0.21"
ipdb = "^0.13.9"
python-decouple = "^3.5"
pre-commit = "^2.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
